Update driver.py
Add a few rooms to test. 
Add the testNameOfUser option to add in the name of who is responsible for the reservation (student or T.A can also do this instead of using professors name for flexibility). 
Add the testTimeNeeded option for a professor to indicate how much time they need for a room to be in use.